requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asn1crypto==1.5.1
attrs==25.4.0
blinker==1.9.0
boto3==1.42.70
botocore==1.42.70
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
cloudpickle==3.1.1
cryptography==46.0.5
exceptiongroup==1.3.1
fastapi==0.128.8
filelock==3.19.1
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.18.0
numpy==2.0.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
protobuf==6.33.5
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyJWT==2.12.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
s3transfer==0.16.0
six==1.17.0
smmap==5.0.3
snowflake-connector-python==4.3.0
snowflake-snowpark-python==1.47.0
sortedcontainers==2.4.0
starlette==0.49.3
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tomlkit==0.14.0
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==1.26.20
uvicorn==0.39.0
